requirements.txt
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
click==8.1.3
cssmin==0.2.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
minify_html==0.10.8
passlib==1.7.4
Pillow==9.4.0
PyJWT==2.6.0
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
redis==4.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
uvicorn==0.20.0
uvloop==0.17.0
validate-email==1.3
watchfiles==0.18.1
webassets==2.0
websockets==10.4
